fix: add domain entities to calendar_import feature
- Extract CalendarImportState and related classes to domain layer
- Update Cubit to import and export the new domain entities
- Update state tests to use the new domain import
- Improve architectural separation in calendar_import feature
