+++ lib/features/calendar_import/application/calendar_import_cubit.dart
@@ -3,43 +3,11 @@ import 'package:injectable/injectable.dart';
 
 import '../../appointments/domain/entities/appointment.dart';
 import '../domain/entities/calendar_event.dart';
+import '../domain/entities/calendar_import_state.dart';
 import '../domain/repositories/calendar_repository.dart';
 import '../domain/usecases/import_calendar_usecase.dart';
 
-// ---------------------------------------------------------------------------
-// State
-// ---------------------------------------------------------------------------
-
-abstract class CalendarImportState {
-  const CalendarImportState();
-}
-
-class CalendarImportInitial extends CalendarImportState {
-  const CalendarImportInitial();
-}
-
-class CalendarImportLoading extends CalendarImportState {
-  const CalendarImportLoading();
-}
-
-class CalendarImportLoaded extends CalendarImportState {
-  final List<Appointment> foundAppointments;
-  const CalendarImportLoaded(this.foundAppointments);
-}
-
-class CalendarImportSuccess extends CalendarImportState {
-  final int importedCount;
-  const CalendarImportSuccess(this.importedCount);
-}
-
-class CalendarImportError extends CalendarImportState {
-  final String message;
-  const CalendarImportError(this.message);
-}
-
-class CalendarImportPermissionDenied extends CalendarImportState {
-  const CalendarImportPermissionDenied();
-}
+export '../domain/entities/calendar_import_state.dart';
 
 // ---------------------------------------------------------------------------
 // Cubit
